pyproject.toml
[tool.poetry]
name = "manage_minyan"
version = "0.1.0"
description = "Auto generate weekly emails for Sabbath Prayer group"
authors = ["Daniel Landesman"]

[tool.poetry.dependencies]
python = "3.7.4"
google-api-python-client = "^1.7.11"
google-auth-oauthlib = "^0.4.1"
html2text = "^2020.1.16"
pyyaml = "^5.3"
python-dotenv = "^0.10.5"
python-crontab = "^2.4.0"
#pyyaml = "^3.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
